+++ blocklets/ai-studio/src/pages/explore/app.tsx
@@ -1,8 +1,8 @@
 import ApplicationHeader from '@app/components/application/ApplicationHeader';
 import { getErrorMessage } from '@app/libs/api';
 import { agentViewTheme } from '@app/theme/agent-view-theme';
+import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
 import Result from '@arcblock/ux/lib/Result';
-import Toast from '@arcblock/ux/lib/Toast';
 import { getAgentByDeploymentId } from '@blocklet/aigne-sdk/api/agent';
 import AgentView from '@blocklet/aigne-sdk/components/AgentView';
 import { Box, CircularProgress, Stack, ThemeProvider } from '@mui/material';
@@ -13,13 +13,12 @@ import { MakeYoursButton, ShareButton } from './button';
 
 export default function AppPage() {
   const { appId } = useParams();
+  const { t } = useLocaleContext();
   if (!appId) throw new Error('Missing required param `appId`');
 
-  const { data, loading, error } = useRequest(() => getAgentByDeploymentId({ deploymentId: appId, working: true }), {
-    onError: (error) => {
-      Toast.error(getErrorMessage(error));
-    },
-  });
+  const { data, loading, error } = useRequest(() => getAgentByDeploymentId({ deploymentId: appId, working: true }));
+
+  const isNoSuchEntryAgentError = (error as any)?.response?.data?.error?.type === 'NoSuchEntryAgentError';
 
   return (
     <ThemeProvider theme={agentViewTheme}>
@@ -32,7 +31,7 @@ export default function AppPage() {
             <ShareButton deployment={data.deployment} project={data.project} />
           </Stack>
 
-          <AgentView aid={data?.identity?.aid} working />
+          <AgentView aid={data?.identity?.aid} />
         </>
       ) : loading ? (
         <Box textAlign="center" my={10}>
@@ -42,7 +41,8 @@ export default function AppPage() {
         <Box
           component={Result}
           status={(error as any)?.response?.status || 500}
-          description={error?.message}
+          title={isNoSuchEntryAgentError ? t('noEntryAgent') : ''}
+          description={isNoSuchEntryAgentError ? t('noEntryAgentDescription') : getErrorMessage(error)}
           sx={{ bgcolor: 'transparent' }}
         />
       )}