+++ blocklets/ai-studio/src/pages/project/deployment-action.tsx
@@ -36,6 +36,7 @@ import {
   Stack,
   SxProps,
   Theme,
+  Tooltip,
   Typography,
   useMediaQuery,
 } from '@mui/material';
@@ -48,6 +49,7 @@ import { Deployment, createDeployment, getDeploymentByProjectId, updateDeploymen
 import PublishView from './publish-view';
 import PublishButton from './publish/publish-button';
 import { saveButtonState } from './state';
+import { useProjectStore } from './yjs-state';
 
 export default function DeploymentAction() {
   const deploymentPopperState = usePopupState({ variant: 'popper', popupId: 'deployment' });
@@ -162,6 +164,8 @@ function DeployApp({
   const isAdmin = useIsAdmin();
   const { t } = useLocaleContext();
   const [visibility, setVisibility] = useState<'public' | 'private'>('public');
+  const { store, config } = useProjectStore(projectId, projectRef);
+  const hasEntry = store.files[config?.entry!];
 
   const handleVisibilityChange = (event: any) => {
     setVisibility(event.target.value);
@@ -259,9 +263,17 @@ function DeployApp({
             <PublishButton />
           </Box>
 
-          <LoadingButton variant="contained" onClick={onSubmit} data-testid="add-deploy-button">
-            {t('deploy')}
-          </LoadingButton>
+          <Tooltip title={!hasEntry ? `${t('noEntryAgent')}, ${t('noEntryAgentDescription')}` : ''}>
+            <span>
+              <LoadingButton
+                variant="contained"
+                onClick={onSubmit}
+                disabled={!hasEntry}
+                data-testid="add-deploy-button">
+                {t('deploy')}
+              </LoadingButton>
+            </span>
+          </Tooltip>
         </Stack>
       </Box>
     </Box>
@@ -271,6 +283,8 @@ function DeployApp({
 function UpdateApp({ id, data, run, sx }: { id: string; data: Deployment; run: () => void; sx?: SxProps }) {
   const { t } = useLocaleContext();
   const isAdmin = useIsAdmin();
+  const { store, config } = useProjectStore(data.projectId, data.projectRef);
+  const hasEntry = store.files[config?.entry!];
 
   const [visibility, setVisibility] = useState<'public' | 'private'>(data.access || 'public');
   const handleVisibilityChange = (event: any) => setVisibility(event.target.value);
@@ -431,9 +445,17 @@ function UpdateApp({ id, data, run, sx }: { id: string; data: Deployment; run: (
             <PublishButton />
           </Box>
 
-          <LoadingButton variant="contained" onClick={onSubmit} data-testid="update-deploy-button">
-            {t('update')}
-          </LoadingButton>
+          <Tooltip title={!hasEntry ? `${t('noEntryAgent')},${t('noEntryAgentDescription')}` : ''}>
+            <span>
+              <LoadingButton
+                variant="contained"
+                onClick={onSubmit}
+                disabled={!hasEntry}
+                data-testid="update-deploy-button">
+                {t('update')}
+              </LoadingButton>
+            </span>
+          </Tooltip>
         </Stack>
       </Box>
     </Stack>